backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.3",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "pillow>=12.1.1",
    "psycopg2>=2.9.11",
    "redis>=7.3.0",
]

[dependency-groups]
dev = ["ruff>=0.15.6"]


[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle
    "F",  # pyflakes
    "I",  # isort
    "UP", # pyupgrade
    "B",  # bugbear
]

ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"migrations/*.py" = ["E501", "F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
